Rename AbstractModule object to Module object
