+++ dl/src/main/scala/com/intel/analytics/bigdl/nn/Module.scala
@@ -1,0 +1,83 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.intel.analytics.bigdl.nn
+
+import com.intel.analytics.bigdl.nn.abstractnn.Activity
+import com.intel.analytics.bigdl.nn.abstractnn.AbstractModule
+import com.intel.analytics.bigdl.tensor.Tensor
+import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
+import com.intel.analytics.bigdl.utils.File
+
+import scala.reflect.ClassTag
+
+object Module {
+  def load[A <: Activity: ClassTag, B <: Activity: ClassTag,
+  @specialized(Float, Double) T: ClassTag](path : String) : AbstractModule[A, B, T] = {
+    File.load[AbstractModule[A, B, T]](path)
+  }
+
+  def flatten[@specialized(Float, Double) T: ClassTag](parameters: Array[Tensor[T]])(
+    implicit ev: TensorNumeric[T]): Tensor[T] = {
+    val compactedTensor = isCompact(parameters)
+    if (compactedTensor != null) {
+      return compactedTensor
+    }
+    var i = 0
+    var length = 0
+    while (i < parameters.length) {
+      require(parameters(i).isContiguous())
+      length += parameters(i).nElement()
+      i += 1
+    }
+
+    val result = Tensor[T](length)
+    val resultStorage = result.storage()
+
+    i = 0
+    var offset = 0
+    while (i < parameters.length) {
+      System.arraycopy(parameters(i).storage().array(), parameters(i).storageOffset() - 1,
+        resultStorage.array(), offset, parameters(i).nElement())
+      parameters(i).set(resultStorage, offset + 1, parameters(i).size(), parameters(i).stride())
+      offset += parameters(i).nElement()
+      i += 1
+    }
+
+    result
+  }
+
+  def isCompact[@specialized(Float, Double) T: ClassTag](paramters: Array[Tensor[T]])(
+    implicit ev: TensorNumeric[T]): Tensor[T] = {
+    require(paramters.length > 0)
+    var i = 1
+    val storage = paramters(0).storage()
+    var length = paramters(0).nElement()
+    while (i < paramters.length) {
+      if (!storage.eq(paramters(i).storage())) {
+        return null
+      }
+      length += paramters(i).nElement()
+      i += 1
+    }
+
+    if (length != storage.array().length) {
+      return null
+    }
+
+    return Tensor(storage)
+  }
+}
